GroundedPredicateNode lib: support is added
